Move .toggleLineComments method from TokenizedBuffer to TextEditor
